requirements.txt
flask
flask-restplus
pdfkit
qrcode[pil]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
